[2019 - Day 07] Converted Computer to take Longs instead Ints

+++ src/main/kotlin/be/edu/adventofcode/y2019/day05/Day05.kt
@@ -4,5 +4,5 @@ import be.edu.adventofcode.Text
 import be.edu.adventofcode.y2019.intcode.Computer
 
 class Day05 {
-    fun bothParts(input: Text, inputInstruction: Int): Int = Computer().execute(input.get(), inputInstruction)
+    fun bothParts(input: Text, inputInstruction: Long): Long = Computer().execute(input.get(), inputInstruction)
 }
